+++ src/driver/main.c
@@ -1,4 +1,6 @@
-#include "dream.h"
+#include "../lexer/lexer.h"
+#include "../parser/parser.h"
+#include "../codegen/codegen.h"
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>